% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{dh_app}
\alias{dh_app}
\title{A Shiny App Walkthrough to Read Colour Hex Codes}
\usage{
dh_app()
}
\value{
Opens a Shiny app.
}
\description{
Rather than manually typing functions of {dehex} into your console, you can
instead run this Shiny app to be guided through the process of learning to
read a random colour hex code.
}
\examples{
\dontrun{dh_app()}
}
