% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehex.R
\name{dh_graph}
\alias{dh_graph}
\title{Print a Hex Shortcode as an RGB graph}
\usage{
dh_graph(hex_short, text = NULL)
}
\arguments{
\item{hex_short}{Character. A valid hex-colour shortcode starting with a
hash mark (#) and followed by three alphanumeric characters, which must
take the values 0 to 9 or A to F (case insensitive).}

\item{text}{Character. An optional string to place above the plot. If NULL
(default), then the shortcode will be automatically selected.}
}
\value{
Nothing. Prints to the console with cat() a bar chart.
}
\description{
Takes a three-character hex colour shortcode and prints to the console a bar
chart of the red (R), green (G) and blue (B) values using unicode blocks.
This provides a quick visual assessment of the relative amounts of each
colour expressed by the hex shortcode.
}
\details{
The amount of red (R), green (G) and blue (B) is calculated on the
basis that hex shortcodes contain one character for each colour. Since
they're hexadecimal, they can take one of 16 possible values: 0 to 9 and
then alphabetic characters from A to F (i.e. 10 to 15). That means that
a shortcode of '#18F' has relative RGB values of 1, 8 and 16.
}
\examples{
dh_graph("#D83")
}
