% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehex.R
\name{dh_hue}
\alias{dh_hue}
\title{Print RGB Graphs to Assess Hue Similarity}
\usage{
dh_hue(hex_short = NULL)
}
\arguments{
\item{hex_short}{Character. A valid hex-colour shortcode starting with a
hash mark (#) and followed by three alphanumeric characters, which must
take the values 0 to 9 or A to F (case insensitive). If NULL (default),
all 16 graphs will be printed along with their common name.}
}
\value{
Nothing. Prints to the console 12 RGB bar charts (if hex_short =
NULL), or prints the RGB graph for the provided shortcode, plus the
RGB graph for the (automatically-selected) 'most similar' hue.
}
\description{
Print RGB graphs for the 12 primary, secondary and tertiary colours. Use the
shape of their distributions to assess the 'most similar hue' to a given hex
shortcode. Can also return the 'answer' for the hue that's 'closest' to your
provided hex shortcode.
}
\examples{
dh_hue("#F73")
}
