% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehex.R
\name{dh_light}
\alias{dh_light}
\title{Print RGB Bar Graphs to Demonstrate Light Profiles}
\usage{
dh_light(light = TRUE)
}
\value{
Nothing. Prints bar charts of RGB values to the console, made with
unicode blocks.
}
\description{
Print the primary, secondary and tertiary colours in the HSL system as bar
charts of the RGB values that they're composed of. Intended for use as a
reference to assess which RGB profile most closely represents a user's hex
shortcode.
}
\details{
The values are light (i.e. high mean value of RGB), middle and dark
(i.e. small range in RGB values).
}
\examples{
dh_hue()
}
