% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehex.R
\name{dh_random}
\alias{dh_random}
\title{Generate a Random Hex Colour Code}
\usage{
dh_random(shorten = FALSE)
}
\arguments{
\item{shorten}{Logical. Shorten six-character hex colour code to three
characters? The three-character hex code is what is used to interpret
the colour in the \{dehex\} 'training' context.}
}
\value{
Character string of length one. A three- or six-character hex code
preceded by a hash mark, '#'.
}
\description{
The purpose of \{dehex\} is to train how to 'read' the colour from a hex code.
This function generates a random hex code for you so you can test your skills.
}
\examples{
dh_random(TRUE)
}
