% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehex.R
\name{dh_sat}
\alias{dh_sat}
\title{Print RGB Bar Graphs to Demonstrate Saturation Profiles}
\usage{
dh_sat(sat = TRUE)
}
\value{
Nothing. Prints bar charts of RGB values to the console, made with
unicode blocks.
}
\description{
Print bar charts of RGB values that represent simplified groupings of
saturation levels in the HSL system. Intended for use as a reference to
assess which RGB profile most closely represents a user's hex shortcode.
}
\details{
The values are saturated (i.e. the largest range in RGB values)
washed, muted and grey (i.e. no difference between RGB values).
}
\examples{
dh_hue()
}
