% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehex.R
\name{dh_shorten}
\alias{dh_shorten}
\title{Convert a Full Hex Code to Shorthand Code}
\usage{
dh_shorten(hex_code)
}
\arguments{
\item{hex}{Character. A valid hex colour code starting with a hash mark (#).
Characters must take the values 0 to 9 or A to F (case insensitive).}
}
\value{
Character. A shortened hex colour code: a hash mark followed by three
characters (0 to 9, A to F).
}
\description{
Takes a hex code like '#C0FFEE' and simplifies to '#CFE' (i.e. the first
character from each character pair).
}
\examples{
dh_shorten("#C0FFEE")
}
