% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehex.R
\name{dh_solve}
\alias{dh_solve}
\title{Describe in Words a Colour from its Hex Code}
\usage{
dh_solve(hex_code)
}
\arguments{
\item{hex_code}{Character. A valid hex colour code starting with a hash mark
(#). Characters must take the values 0 to 9 or A to F (case insensitive).}
}
\value{
Character. A string describing the input hex colour in words, like
'middle saturated azure'.
}
\description{
Describe in Words a Colour from its Hex Code
}
\examples{
dh_solve("#08F")
}
