% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehex-package.R
\docType{package}
\name{dehex-package}
\alias{dehex}
\alias{dehex-package}
\title{dehex: Learn to Roughly Assess Colour Hex Codes By Eye}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Follow David DeSandro's method to quickly generate a rough colour
    name (e.g. 'light washed yellow') by looking only at its hex code (e.g. 
    #F2D359). Contains standalone functions and a Shiny app.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/dehex}
  \item Report bugs at \url{https://github.com/matt-dray/dehex/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mattdray@gmail.com}

}
\keyword{internal}
