% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehex-package.R
\docType{package}
\name{dehex-package}
\alias{dehex}
\alias{dehex-package}
\title{dehex: Learn to Quickly Assess Hex Code Colours}
\description{
Follow David DeSanto's method to quickly generate a rough colour
    name by looking only at its hex code.
}
\author{
\strong{Maintainer}: Matt Dray \email{mattdray@gmail.com}

}
\keyword{internal}
