% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehex.R
\name{dh_solve}
\alias{dh_solve}
\title{Get a Name for a Hex Code With Optional Diagnostic Plots and Swatch}
\usage{
dh_solve(hex_code, swatch = FALSE, graphs = FALSE)
}
\arguments{
\item{hex_code}{Character. A valid hex colour code starting with a hash mark
(#). Characters must take the values 0 to 9 or A to F (case insensitive).}

\item{swatch}{Logical. Print to a graphical device a plot of the colour
represented by (three-digit) hex code? Defaults to FALSE.}

\item{graphs}{Logical. Do you want to print the result and associated hue,
saturation and lightness bar charts to the console? Defaults to FALSE.}
}
\value{
A character string. Optionally some console output and a plot.
}
\description{
Convert a colour hex code to an English string that roughly describes its
colour in terms of hue, saturation and lightness, like 'dark saturated
azure'. Optionally plot a block with the colour that the hex code encodes.
Optionally print to the console the hue, saturation and lightness
graphs that best approximate that hex code.
}
\examples{
dh_solve("#08F", graphs = FALSE, swatch = FALSE)
}
