% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehex.R
\name{dh_swatch}
\alias{dh_swatch}
\title{Plot the Colour of a Short Hex Code}
\usage{
dh_swatch(hex_code)
}
\arguments{
\item{hex_code}{Character. A valid hex colour code starting with a hash mark
(#). Characters must take the values 0 to 9 or A to F (case insensitive).}
}
\value{
A {grid} graphics plot.
}
\description{
Check the colour of an input hex colour code by plotting it.
}
\examples{
dh_swatch("#F14362")
}
